Add --bub-check --edge-to-seq and --neighbour-steps options.
These are mainly to be used in conjunction with -c0 which removes the
minimum copy-number of 1 setting.  However this often truncates graph
paths, so the options are designed to add back potentially useful
nodes even if coverage data is lacking.

Bub-check rescues copy-number on bubbles.  A->{P,Q}->B will for nodes
A and B with coverage but none of *either* P *nor* Q will add P and
Q.  If one of P or Q is covered then a path exists and the other node
is deleted.

Edge-to-seq is for use with kmer2node4 without the -U option.  This
can add edge information on nodes transitioning to or from duplicated
kmers, but it doesn't add to depth as it's sometimes erroneous.  This
rescues some of that depth by adding minimum copy-number estimation
based on arc coverage too.  (It's not a  particularly promising change
however.)

Neighbour-steps N iteratively increases minimum copy number to 1 for
any node found to be within N steps of a node with copy_number >= 1.
It's like -c1 used to be, but still excluding more distant parts of
the graph.